Added Kolto, Lesser Kolto for a Bacta alternative
Kolto is an early medical substance used by Both the Jedi and Sith during Kotor I-II, it is still used in the Clone Wars and GCW but its extremely expensive which is why Bacta is used more. Currently have three new chemicals added to the game. (Lesser-Kolto, Kolto, and Bacta) Lesser-Kolto does not exist on the Wiki but it allows people to not instantly have Kolto, it is in the chem machine for now.